Divide methods into classes, updated logic.
