+++ 3614/1/gems.yml
@@ -4,9 +4,9 @@ gems:
   - pry-byebug
   - letter_opener
   - kaminari
-  - paranoia
-  - aasm
-  - activerecord-import
-  - rollout
-  - grape
-  - rake
+ # - paranoia
+ # - aasm
+ # - activerecord-import
+ # - rollout
+ # - grape
+ # - rake
